SNOW-986492-console-output-tools: move make fake app fixture to upper level
